awesomeBot/requirements.txt
aiocache==0.11.1
aiocqhttp==1.4.3
aiofiles==0.8.0
anyio==3.6.1
blinker==1.5
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==0.6.2
nonebot==1.9.1
priority==2.0.0
PyYAML==6.0
quart==0.17.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.2.0
toml==0.10.2
urllib3==1.26.10
Werkzeug==2.1.2
wsproto==1.1.0
